Added goomph to the IDE tools.
Goomph downloads an Eclipse IDE and performs all setup actions necessary to create a complete workspace with the required plugins and settings.  Thanks to an Eclipse feature called "bundle pooling", these installations are very efficient with disk space.  All the plugins for multiple installations are installed to one place (the "bundle pool"), and the installations just point to those jars.